requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
sqlalchemy==1.4.31
pydantic
pyjwt
bcrypt
python-multipart
azure-storage-blob